Implemented authentication successfully
