+++ server/routes/user.routes.js
@@ -2,14 +2,19 @@ import express from 'express';
 import {
   getAllUsers,
   getUserById,
-  createUser,
   updateUser,
   deleteUser,
-  otpverify,
-  loginUser
 } from '../controllers/user.controller.js';
 
+import {
+  createUser,
+  otpverify,
+  loginUser,
+} from '../controllers/user.auth.js';
+
 const router = express.Router();
+
+
 
 router.get('/users', getAllUsers);
 router.get('/user/:id', getUserById);
@@ -19,3 +24,5 @@ router.post('/otp-verfy',otpverify)
 router.put('/user/:id', updateUser);
 router.delete('/user/:id', deleteUser);
 export default router;
+
+
