+++ server/src/models/user.model.js
@@ -1,16 +1,4 @@
-// import mongoose from "mongoose";
 
-// const UserSchema = new mongoose.Schema(
-//   {
-//     email:      { type: String, required: true, unique: true },
-//     password:   { type: String, required: true },      
-//     name:       { type: String, required: true },
-//     phone:      String,
-//   },
-//   { timestamps: true }
-// );
-
-// export default mongoose.model("User", UserSchema);
 
 import mongoose from "mongoose";
